\name{ci}
\alias{ci}
\title{
Confidence interval
}
\description{
Get back the confidence interval after a bootstrap simulation.
}
\usage{
ci(x, conf.level = 0.95)
}
\arguments{
  \item{x}{
numeric vector.
}
  \item{conf.level}{
confidence level.
}
}
\author{
Maxime Hervé <mx.herve@gmail.com>
}