\name{cond.multinom}
\alias{cond.multinom}
\title{
Condition number of the Hessian matrix of a multinomial log-linear model
}
\description{
Computes the condition number of the Hessian matrix of a model fitted with \code{\link[nnet]{multinom}}.
}
\usage{
cond.multinom(model)
}
\arguments{
  \item{model}{
object of class \code{"multinom"}.
}
}
\author{
Maxime Hervé <mx.herve@gmail.com>
}