\name{print.byf.test}
\alias{print.byf.test}
\title{
Print method
}
\description{
Print method for multiple tests (one per level of a factor).
}
\usage{
\method{print}{byf.test}(x, ...)
}
\arguments{
  \item{x}{
object of class \code{byf.test}.
}
  \item{...}{
further arguments.
}
}
\author{
Maxime Hervé <mx.herve@gmail.com>
}
