\name{print.RV.multcomp}
\alias{print.RV.multcomp}
\title{
Print method
}
\description{
Print method for pairwise comparisons tables.
}
\usage{
\method{print}{RV.multcomp}(x, digits = 4, ...)
}
\arguments{
  \item{x}{
object of class \code{RV.multcomp}.
}
  \item{digits}{
minimal number of significant digits.
}
  \item{...}{
further arguments.
}
}
\author{
Maxime Hervé <mx.herve@gmail.com>
}
