\name{print.RVtest}
\alias{print.RVtest}
\title{
Print method
}
\description{
Print method for a global test followed by pairwise comparisons tables or evaluation of a common coefficient.
}
\usage{
\method{print}{RVtest}(x, digits = 4, quote = TRUE, prefix = "", ...)
}
\arguments{
  \item{x}{
object of class \code{RVtest}.
}
  \item{digits}{
minimal number of significant digits.
}
  \item{quote}{
logical, indicating whether or not strings should be printed with surrounding quotes.
}
  \item{prefix}{
.
}
  \item{...}{
further arguments.
}
}
\author{
Maxime Hervé <mx.herve@gmail.com>
}
