\name{RVAideMemoire-package}
\alias{RVAideMemoire-package}
\alias{RVAideMemoire}
\docType{package}
\title{
Fonctions diverses accompagnant l'ouvrage 'Aide-memoire de statistique appliquee a la biologie'
}
\description{
Ce package contient des fonctions à buts diverses, plus ou moins complexes, écrites pour simplifier la vie des débutants en R.  Elles sont présentées dans l'ouvrage \emph{"Aide-mémoire de statistique appliquée à la biologie"} du même auteur, disponible sur le CRAN.
}
\details{
\tabular{ll}{
Package: \tab RVAideMemoire\cr
Type: \tab Package\cr
Version: \tab 0.9.1\cr
Date: \tab 2011-03-02\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Maxime Hervé

Maintainer: Maxime Hervé <maxime.herve@laposte.net>
}
\references{
Document : "Aide-memoire de statistique appliquee a la biologie - Construire son etude et analyser les resutats a l'aide du logiciel R"
}
\seealso{
\code{\link[ade4:ade4-package]{ade4}}, \code{\link[boot]{boot}}, \code{\link[mgcv:mgcv-package]{mgcv}}
}
