\name{RVAideMemoire-package}
\alias{RVAideMemoire-package}
\alias{RVAideMemoire}
\docType{package}
\title{
Testing and Plotting Procedures for Biostatistics
}
\description{
Contains miscellaneous functions useful in biostatistics, mostly univariate and multivariate testing procedures with a special emphasis on permutation tests. Many functions intend to simplify user's life by shortening existing procedures or by implementing plotting functions that can be used with as many methods from different packages as possible.
}
\details{
\tabular{ll}{
Package: \tab RVAideMemoire\cr
Type: \tab Package\cr
Version: \tab 0.9-81-2\cr
Date: \tab 2022-02-21\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Maxime Hervé

Maintainer: Maxime Hervé <maxime.herve@univ-rennes1.fr>
}
\references{
Document : "Aide-memoire de statistique appliquee a la biologie - Construire son etude et analyser les resutats a l'aide du logiciel R" (available on CRAN)
}