\name{RVAideMemoire-package}
\alias{RVAideMemoire-package}
\alias{RVAideMemoire}
\docType{package}
\title{
Diverse basic statistical and graphical functions
}
\description{
This package contains diverse more or less complicated functions, written to simplify beginners' (but not only) life: simplifications of existing functions, basic but not implemented tests, easy-to-use tools... All functions are presented in the French book 'Aide-memoire de statistique appliquee a la biologie', written by the same author and available on CRAN.
}
\details{
\tabular{ll}{
Package: \tab RVAideMemoire\cr
Type: \tab Package\cr
Version: \tab 0.9-10\cr
Date: \tab 2012-06-16\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Maxime Hervé

Maintainer: Maxime Hervé <mx.herve@gmail.com>
}
\references{
Document : "Aide-memoire de statistique appliquee a la biologie - Construire son etude et analyser les resutats a l'aide du logiciel R" (available on CRAN)
}